fix ndigits for bases > 62, and make it way faster
The C library function `sizeinbase` is valid only for bases <= 62.
So use Julia's implementation, which is valid for all bases,
and happens to be quite faster, e.g. by 5x or 10x
(the reason is in its faster workaround to the fact that even
for correct bases, `sizeinbase` can return a value which is "one too big".)