set `subfields` and `overfields` ...
... also for `FqDefaultFiniteField` and `FqFiniteField`,
not only for `FqNmodFiniteField`.

This seems to be not controversial.
The fix does not really help much in my situation (`embed` for finite fields)
since afterwards I run into the error that `embed_matrices` is not
available for `FqFiniteField`.
The available methods are in the file `fq_nmod_embed.jl`,
thus I guess there should be analogous files `fq_default_embed.jl`,
`q_embed.jl` or so?