Avoid String->Symbol conversion
This also includes changing various methods to accept inputs of
type `VarName` instead of `String` for specifying variable names,
so that users can specify variable names as symbols (and thus avoid
conversion). It also improves the user experience as it means more
places accept both strings and symbols (and chars).