move code from `Oscar.jl/experimental/GModule/Misc.jl` here
The idea is to add this code to Nemo.jl,
and then to remove it from Oscar.jl.

Concerning the usefulness of this code (and tests to be added),
I am not sure which vectors shall be supported by the new methods;
vectors of integers are already addressed by other methods.