RFC: redesign of the flint wrapper
The FLINT wrapper is inefficient when accessing elements of matrices
or polynomials (or anything else nested). The problem is that
to acesss `M[i, j]` we first create a new `fmpz` and then `fmpz_set`
the elements. As a consequence, we now have a lot of `mat_entry_ptr`
functions, which are very (very) unsafe and plenty of `ZZRingElemOrPtr`
occurences.

I had many discussions with @fieker and Dan over the last few years
and here is one altenative design which is distilled from these
discussions:

```
struct ZZRingElemDR
  ptr::Ptr{ZZRingElem}
  data #= can be anything where ptr points to (fmpz, fmpz_mat, fmpz_poly, ...) =#
end
```

("DR" stands for "done right", a tribute to Dan.)

The idea is to have an (immutable) layer on top of the raw `fmpz`,
which holds a pointer to an `fmpz` together with "data". The "data"
part is only there to keep the memory alive for the sake of the GC.
It can be a `ZZRingElem`, or a `ZZMatrix` or a `ZZPolyRingElem`, and so
on.

Together with some `Base.cconvert`, it is relatively easy to make this
fly, see the diff:

```
julia> a = Nemo.ZZRingElemDR(ZZ(2))
2

julia> -a
-2

julia> a = Nemo.ZZRingElemDR(ZZ[1 2; 1 1], 1, 2)
2

julia> is_unit(a)
false

julia> Zx, x = ZZ["x"]; f = x^3 + 213*x + 1;

julia> -Nemo.ZZRingElemDR(f, 1)
-213
```

The important point is that this is allocation free:

```
julia> M = ZZ[1 2; 3 4];

julia> @ballocated getindex(M, 1, 1)
16

julia> @ballocated Nemo.ZZRingElemDR(M, 1, 2)
0
```

One drawback: This increases size of FLINT types. We add 16 bytes per
object. For small `fmpz`, this makes them 3x larger.